Refactor ListHosts to new endpoint pattern
