+++ server/service/endpoint_hosts.go
@@ -85,42 +85,6 @@ func makeHostByIdentifierEndpoint(svc fleet.Service) endpoint.Endpoint {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// List Hosts
-////////////////////////////////////////////////////////////////////////////////
-
-type listHostsRequest struct {
-	ListOptions fleet.HostListOptions
-}
-
-type listHostsResponse struct {
-	Hosts []HostResponse `json:"hosts"`
-	Err   error          `json:"error,omitempty"`
-}
-
-func (r listHostsResponse) error() error { return r.Err }
-
-func makeListHostsEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(listHostsRequest)
-		hosts, err := svc.ListHosts(ctx, req.ListOptions)
-		if err != nil {
-			return listHostsResponse{Err: err}, nil
-		}
-
-		hostResponses := make([]HostResponse, len(hosts))
-		for i, host := range hosts {
-			h, err := hostResponseForHost(ctx, svc, host)
-			if err != nil {
-				return listHostsResponse{Err: err}, nil
-			}
-
-			hostResponses[i] = *h
-		}
-		return listHostsResponse{Hosts: hostResponses}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Get Host Summary
 ////////////////////////////////////////////////////////////////////////////////
